requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
langchain-upstage
python-dotenv

